+++ Content.Shared/CrystallPunk/LockKey/Components/CPKeyComponent.cs
@@ -8,7 +8,7 @@ namespace Content.Shared.CrystallPunk.LockKey;
 /// <summary>
 /// a key component that can be used to unlock and lock locks from CPLockComponent
 /// </summary>
-[RegisterComponent, Access(typeof(SharedCPLockKeySystem))]
+[RegisterComponent]
 public sealed partial class CPKeyComponent : Component
 {
     [DataField]
@@ -18,5 +18,5 @@ public sealed partial class CPKeyComponent : Component
     /// If not null, automatically generates a key for the specified category on initialization. This ensures that the lock will be opened with a key of the same category.
     /// </summary>
     [DataField]
-    public ProtoId<CPLockCategoryPrototype>? AutoGenerateKey = null;
+    public ProtoId<CPLockCategoryPrototype>? AutoGenerateShape = null;
 }
