+++ Content.Shared/CrystallPunk/LockKey/Components/CPLockComponent.cs
@@ -1,4 +1,5 @@
 using Content.Shared.CrystallPunk.LockKey;
+using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 
 namespace Content.Shared.CrystallPunk.LockKey;
@@ -6,7 +7,7 @@ namespace Content.Shared.CrystallPunk.LockKey;
 /// <summary>
 /// A component of a lock that stores its keyhole shape, complexity, and current state.
 /// </summary>
-[RegisterComponent, Access(typeof(SharedCPLockKeySystem))]
+[RegisterComponent]
 public sealed partial class CPLockComponent : Component
 {
     [DataField]
@@ -15,12 +16,9 @@ public sealed partial class CPLockComponent : Component
     [DataField]
     public float LockPickBreakChance = 0.3f;
 
-    [DataField]
-    public bool Locked = false;
-
     /// <summary>
     /// If not null, automatically generates a lock for the specified category on initialization. This ensures that the lock will be opened with a key of the same category.
     /// </summary>
     [DataField]
-    public ProtoId<CPLockCategoryPrototype>? AutoGenerateLock = null;
+    public ProtoId<CPLockCategoryPrototype>? AutoGenerateShape = null;
 }
