+++ Content.Shared/Lock/LockComponent.cs
@@ -71,6 +71,12 @@ public sealed partial class LockComponent : Component
     [DataField]
     [AutoNetworkedField]
     public TimeSpan UnlockTime;
+
+    /// <summary>
+    /// CrystallPunk LockSystem Adapt: we need LockEntity in object, in slotId.
+    /// </summary>
+    [DataField]
+    public string? LockSlotId = string.Empty;
 }
 
 /// <summary>
