+++ Resources/Locale/ru-RU/lock/lock-component.ftl
@@ -1,0 +1,9 @@
+lock-comp-on-examined-is-locked = [color=#a84032]Он заперт[/color].
+lock-comp-on-examined-is-unlocked = [color=#32a834]Он открыт[/color].
+lock-comp-do-lock-success = Вы заблокировали { $entityName }.
+lock-comp-do-unlock-success = Вы разблокировали { $entityName }.
+
+## ToggleLockVerb
+
+toggle-lock-verb-unlock = Разблокировать
+toggle-lock-verb-lock = Заблокировать
