+++ Resources/Prototypes/CrystallPunk/LockCategories/LockCategory.yml
@@ -19,7 +19,7 @@
   name: ключ тест
   components: 
   - type: CPKey
-    autoGenerateKey: Debug
+    autoGenerateShape: Debug
   
 - type: entity
   parent: SurvivalKnife
@@ -27,7 +27,7 @@
   name: ключ тест 2
   components: 
   - type: CPKey
-    autoGenerateKey: Debug2
+    autoGenerateShape: Debug2
   
 - type: entity
   parent: SurvivalKnife
@@ -35,15 +35,17 @@
   name: ключ тест 3
   components: 
   - type: CPKey
-    autoGenerateKey: Debug3
+    autoGenerateShape: Debug3
 
 - type: entity
   parent: WoodDoor
   id: WoodDoorTest
-  name: тестовая дверь закрытая
+  name: тестовая дверь
+  suffix: Замок 1
   description: Тут есть замок?
   components:
-  - type: CPLockSlot
+  - type: Lock
+    locked: true
     lockSlotId: lock_slot
   - type: ItemSlots
     slots: 
@@ -61,13 +63,18 @@
 - type: entity
   parent: WoodDoorTest
   id: WoodDoorTest2
-  name: тестовая дверь открытая
+  name: тестовая дверь
+  suffix: Открыто
   description: Тут есть замок?
   components:
+  - type: Lock
+    locked: false
+    lockSlotId: lock_slot
   - type: ItemSlots
     slots: 
       lock_slot:
         name: cp-lock-slot-component-slot-name-default
+        disableEject: true
         whitelist:
           components:
           - CPLock
@@ -77,7 +84,7 @@
   id: CrateGenericBiosuitTEST
   name: тестовый ящик
   components:
-  - type: CPLockSlot
+  - type: Lock
     lockSlotId: lock_slot
   - type: ItemSlots
     slots: 
@@ -108,19 +115,19 @@
   
 
 - type: entity
-  parent: Paper
+  parent: SurvivalKnife
   id: TestLock1
   name: тестовый первый замок
   components:
   - type: CPLock
-    autoGenerateLock: Debug
+    autoGenerateShape: Debug
     
 - type: entity
-  parent: Paper
+  parent: SurvivalKnife
   id: TestLock2
   name: тестовый второй замок
   components:
   - type: CPLock
-    autoGenerateLock: Debug2
+    autoGenerateShape: Debug2
